+++ src/schema/mysql.mts
@@ -1,0 +1,23 @@
+import { int, mysqlTable, timestamp, varchar } from "drizzle-orm/mysql-core";
+
+export const mysqlHomeAssistantEntity = mysqlTable("HomeAssistantEntity", {
+  app_unique_id: varchar("app_unique_id", { length: 255 }).notNull(),
+  application_name: varchar("application_name", { length: 255 }).notNull(),
+  base_state: varchar("base_state", { length: 1000 }).notNull(),
+  entity_id: varchar("entity_id", { length: 255 }).notNull(),
+  first_observed: timestamp("first_observed").notNull().defaultNow(),
+  id: int("id").primaryKey().autoincrement(),
+  last_modified: varchar("last_modified", { length: 255 }).notNull(),
+  last_reported: varchar("last_reported", { length: 255 }).notNull(),
+  state_json: varchar("state_json", { length: 1000 }).notNull(),
+  unique_id: varchar("unique_id", { length: 255 }).notNull().unique(),
+});
+
+export const mysqlHomeAssistantEntityLocals = mysqlTable("HomeAssistantEntityLocals", {
+  app_unique_id: varchar("app_unique_id", { length: 255 }).notNull(),
+  id: int("id").primaryKey().autoincrement(),
+  key: varchar("key", { length: 255 }).notNull(),
+  last_modified: timestamp("last_modified").notNull().defaultNow(),
+  unique_id: varchar("unique_id", { length: 255 }).notNull(),
+  value_json: varchar("value_json", { length: 1000 }).notNull(),
+});
